+++ code/ui/interaction/button.tsx
@@ -18,7 +18,7 @@ export type ButtonVariant = "link"
 
 export interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
   variant?: ButtonVariant;
-  size?: "xs" | "sm" | "md" | "lg" | "default" | "icon";
+  size?: "xs" | "sm" | "md" | "lg" | "default" | "icon" | "icon-sm" | "icon-md" | "icon-xs";
   leftIcon?: "string" | React.ReactNode;
   rightIcon?: "string" | React.ReactNode;
   iconSize?: "inherit" | "medium" | "small" | "large";
@@ -75,7 +75,7 @@ export default function Button({
     info:
       "text-white bg-blue-500 hover:bg-blue-600 dark:bg-blue-500 dark:hover:bg-blue-600",
     info_banner:
-      "bg-info-background border border-info-border text-info-foreground shadow-xs ",
+      "bg-info-background border border-info-border text-info-foreground shadow-xs hover:bg-info-background-hover",
     warning:
       "bg-amber-300 hover:bg-amber-500/80 dark:bg-amber-800 dark:hover:bg-amber-900/80",
     secondary: "bg-gray-200 text-gray-800 hover:bg-gray-300",
@@ -99,6 +99,9 @@ export default function Button({
     default: "h-9 px-4 py-2 has-[>svg]:px-3",
     lg: "h-10 rounded-md px-6 has-[>svg]:px-4",
     icon: "size-9",
+    "icon-xs": "size-6",
+    "icon-sm": "size-7",
+    "icon-md": "size-8",
   };
 
   // Define spacing between icons based on size
@@ -109,6 +112,9 @@ export default function Button({
     default: "space-x-2",
     lg: "space-x-2",
     icon: "space-x-0",
+    "icon-xs": "space-x-0",
+    "icon-sm": "space-x-0",
+    "icon-md": "space-x-0",
   };
 
   // Define styles for the disabled state